Minor refactoring (moved hpe_proc to the uav_ctl script
