+++ src/pose_estimation/hpe_3d.py
@@ -376,45 +376,6 @@ class HPE2Dto3D():
             except Exception as e:
                 rospy.logwarn("Failed to generate or publish right hand message: {}".format(e))
 
-    def proc_hpe_est(self):
-        # TODO: Everything in this method should be moved to the separate ctl script
-        try:
-            hpe3d_msg = self.get_hpe3d(copy.deepcopy(self.predictions))
-            self.hpe3d_pub.publish(hpe3d_msg)
-
-            # TODO: Get torso coordinate frame [move this to a method]
-            # TODO: Compare this to the online estimation of the HPE by openpose
-            c_d_ls = pointToArray(hpe3d_msg.l_shoulder)
-            c_d_rs = pointToArray(hpe3d_msg.r_shoulder)
-            c_d_t  = pointToArray(hpe3d_msg.neck)
-            c_d_n  = pointToArray(hpe3d_msg.nose)
-            c_d_le = pointToArray(hpe3d_msg.l_elbow)
-            c_d_re = pointToArray(hpe3d_msg.r_elbow)
-            c_d_rw = pointToArray(hpe3d_msg.r_wrist)
-            c_d_lw = pointToArray(hpe3d_msg.l_wrist)
-
-            cD = np.array([create_homogenous_vector(c_d_t),
-                           create_homogenous_vector(c_d_ls), 
-                           create_homogenous_vector(c_d_rs), 
-                           create_homogenous_vector(c_d_le), 
-                           create_homogenous_vector(c_d_re), 
-                           create_homogenous_vector(c_d_lw), 
-                           create_homogenous_vector(c_d_rw)])
-
-            # body in the camera coordinate frame 
-            bRc = np.matmul(get_RotX(np.pi/2), get_RotY(np.pi/2))
-            # thorax in the camera frame --> TODO: Fix transformations
-            T = create_homogenous_matrix(bRc.T, c_d_t)
-            T_inv = np.linalg.inv(T)
-            # This seems like ok transformation for beginning :) 
-            bD = np.matmul(T_inv, cD.T).T
-            self.publishMarkerArray(bD)             
-
-            torso_msg = self.packSimpleTorso3DMsg(bD)
-            self.upper_body_3d_pub.publish(torso_msg)
-        except Exception as e:
-            rospy.logwarn("Failed to generate or publish HPE3d message: {}".format(e))
-
     def packSimpleTorso3DMsg(self, bD):
         msg = TorsoJointPositions()
         msg.header = self.pcl.header
@@ -507,6 +468,7 @@ class HPE2Dto3D():
                 self.debug_print()
 
             self.rate.sleep()
+
 
 def convert_pose_predictions_to_dict(predictions):
